Refactor tax component structure and add new components
Reorganized the project file structure to better separate concerns: the MainInfo component has been moved to a new location. Added new components for Form 1099 and W2, both for display and creation, creating a more streamlined user experience for inputting data. This refactoring helps ensure that our application is easy to maintain and expand. Also, implemented an IncomesForm component for the TaxFiling page.